requirements.txt
langchain>=0.2.0
langgraph>=0.1.0
langchain-groq>=0.1.0
python-dotenv>=1.0.0
rich>=13.0.0
